requirements.txt
art==6.4
colorama==0.4.6
numpy==2.2.2
opencv-python==4.11.0.86
scipy==1.15.1
pillow==11.1.0

